+++ Projects/Network/Sources/API/CategoryAPI.swift
@@ -1,0 +1,16 @@
+//
+//  CategoryAPI.swift
+//  Network
+//
+//  Created by 리아 on 2022/10/21.
+//  Copyright © 2022 zesty. All rights reserved.
+//
+
+import Combine
+import Foundation
+
+public struct CategoryAPI {
+    
+    static let networkService = NetworkService()
+
+}
